Added script to bulk update ObjectID/ExternalId in Workplace
This PowerShell script allows to change the ObjectID/External ID of Workplace users in bulk.
This is particularly useful for situations in which users are getting migrated from one AD to another while thier email does not change, but their AD Object ID does and so the ExternalId in Workplace needs to be synced accordingly.